+++ pilsner/utility.py
@@ -125,6 +125,29 @@ class Utility():
             del subtrie[model.ENTITY_KEY]
             return len(subtrie) + 1, True
 
+    def ignore_node(self, model, label):
+        """Looks up *label* in a given *model* and hooks a special tag to the leaf in case *label* is found.
+        Tagged label will not be racognized by Utility.spot_entities() function.
+
+        Args:
+            Model *model*: Model instance to look up
+            str *label*: string to tag
+        """
+        label_length = len(label)
+        string_so_far = ''
+        character_index = 0
+        for section in model[model.DICTIONARY_KEY]:
+            content = section[model.CONTENT_KEY]
+            for tokenizer_key in content:
+                trie = content[tokenizer_key]
+                for character_index in range(0, label_length):
+                    string_so_far += label[character_index]
+                    if string_so_far in trie:
+                        trie = trie[string_so_far]
+                        string_so_far = ''
+                if character_index == label_length - 1 and model.ENTITY_KEY in trie and string_so_far == '':
+                    trie[model.IGNORE_KEY] = []
+
     def make_recognizer(self, model, filename, specs, word_separator, item_limit, compressed, column_separator, column_enclosure, tokenizer_option):
         """Reads tab-delimited text file, populates dict objects representing tries, and fills database associated with a given Model instance according to provided specs.
         Returns tuple(list *tries*, dict *line_numbers*) where *tries* are populated dicts representing tries, *line_numbers* is dict that maps line numbers from the text file to internally generated entity IDs.
@@ -429,7 +452,7 @@ class Utility():
                 else: # reading entity
                     end_index = current_index
                     character = source_string[current_index]
-                    if character == word_separator and model.ENTITY_KEY in subtrie:
+                    if character == word_separator and model.ENTITY_KEY in subtrie and model.IGNORE_KEY not in subtrie:
                         found_object = self.check_attrs(model, subtrie, model.cursor, include_query, exclude_query, process_exclude, attrs_out_query)
                         if found_object:
                             identified = found_object[model.ENTITY_KEY], found_object[model.ATTRS_KEY]
@@ -442,7 +465,7 @@ class Utility():
                     else:
                         #if everything_or_nothing and current_index == total_length: return []
                         if character == word_separator or current_index == total_length: # - 1:
-                            if model.ENTITY_KEY in subtrie:
+                            if model.ENTITY_KEY in subtrie and model.IGNORE_KEY not in subtrie:
                                 found_object = self.check_attrs(model, subtrie, model.cursor, include_query, exclude_query, process_exclude, attrs_out_query)
                                 if found_object:
                                     identified = found_object[model.ENTITY_KEY], found_object[model.ATTRS_KEY]
@@ -471,7 +494,7 @@ class Utility():
                         start_index = current_index
                         subtrie = trie[model.CONTENT_KEY][normalizer_name]
                 current_index += 1
-            if model.ENTITY_KEY in subtrie:
+            if model.ENTITY_KEY in subtrie and model.IGNORE_KEY not in subtrie:
                 found_object = self.check_attrs(model, subtrie, model.cursor, include_query, exclude_query, process_exclude, attrs_out_query)
                 if found_object:
                     identified = found_object[model.ENTITY_KEY], found_object[model.ATTRS_KEY]